fix tests dependencies for examples
